refactor: all builds use context
